% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{supported_datasets}
\alias{supported_datasets}
\title{Supported Datasets}
\format{
\subsection{\code{supported_datasets}}{

A data frame with 1 row and one column:
\describe{
\item{data}{Data}
\item{dataset_file_name}{Name of file provided by HCUPS}
}
}
}
\source{
Updated with the addition of a new dataset
}
\usage{
supported_datasets
}
\description{
A list datasets that are currently supported by the package
}
\keyword{datasets}
