% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptions.R
\name{descriptions}
\alias{descriptions}
\title{Return a table of variables and their descriptions}
\usage{
descriptions(dataset, year)
}
\arguments{
\item{dataset}{The dataset (currently only supporting NIS).}

\item{year}{The dataset year.}
}
\value{
A tibble consisting of the variable names and their descriptions.
}
\description{
Return a table of variables and their descriptions
}
\examples{
# The descriptions for the first 5 variables
var_desc <- descriptions("nis", 2019)
head(var_desc, 5)
}
