% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nis.R
\name{read_nis}
\alias{read_nis}
\title{Read NIS file}
\usage{
read_nis(dataset, year, import_method = "readr", ...)
}
\arguments{
\item{dataset}{NIS dataset file}

\item{year}{The year of the dataset}

\item{import_method}{The function to import data. The default is readr/}

\item{...}{Additional arguments}
}
\value{
Returns a tibble
}
\description{
Read NIS file
}
\examples{
path <- system.file("extdata", "NIS_2019_test_data.ASC", package = "loadHCUP")
NIS_2019_df <- read_nis(path, 2019)
}
