% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nis.R
\name{read_nis}
\alias{read_nis}
\title{Read NIS file}
\usage{
read_nis(dataset, year, import_method = "readr", ...)
}
\arguments{
\item{dataset}{Path to NIS dataset file}

\item{year}{The year of the dataset}

\item{import_method}{The function to import data. The default is readr/}

\item{...}{Additional arguments}
}
\value{
Returns a tibble
}
\description{
Read NIS file
}
\examples{
\dontrun{
NIS_2019_df <- read_nis("NIS_2019_test_data.ASC", 2019)
}
}
